producer/requirements.txt
confluent-kafka
requests
sseclient-py
